Parse weekdays into datetime relative to today
Webservice action output from Qualtrics provides a day of the week
for scheduling the text messages. Added support for turning the day
of the week (for example, "Wednesday") into a datetime. It is assumed
that messages will not be sent today (for example if today is actually
Tuesday, and the participant selects Tuesday, text messages will be sent
7 days from today).